Open getaddrinfo for public use
