+++ Development/mdns/test/mdns_test.cpp
@@ -415,3 +415,23 @@ BST_TEST_CASE(testMdnsImpl)
     advertiser.close().wait();
     BST_REQUIRE(gate.hasLogMessage("Close"));
 }
+
+////////////////////////////////////////////////////////////////////////////////////////////
+BST_TEST_CASE(testDnsGetAddrInfo)
+{
+    test_gate gate;
+
+    auto cancellation_source = pplx::cancellation_token_source();
+    auto token = cancellation_source.get_token();
+
+    mdns::details::cancellation_guard guard(token);
+
+    std::string ip_address;
+    BST_REQUIRE(mdns_details::getaddrinfo([&](const mdns_details::address_result& result)
+    {
+        ip_address = result.ip_address;
+        return true;
+    }, "google-public-dns-a.google.com", 0, std::chrono::seconds(2), guard.target, gate));
+
+    BST_REQUIRE(ip_address == "8.8.8.8");
+}
